+++ src/App.jsx
@@ -5,7 +5,6 @@ import { Provider } from "react-redux";
 import store from "./redux/store";
 import Navbar from "./components/Navbar";
 import ProductosDetail from './components/ProductosDetail';
-import CreateClient from './components/CreateClient';
 import { PrivateRoutes, PublicRoutes } from "./models/routes";
 import AuthGuard from "./guards/auth.guard";
 import RoutesWhitNotFound from "./utilities/routesWhitNotFound.utility";
@@ -14,15 +13,14 @@ import { FiltersProvider } from "./components/context/filters";
 import { CartProvider } from "./components/context/CartContext";
 import Loading from "./components/Loading";
 
-const Login = lazy(() => import('./components/Login')); //ver si esto de Login/Login esta bien
-const Private = lazy(() => import('./Private/Private')); //ver si esto de private/Private esta bien minuscula mayuscula
+const Login = lazy(() => import('./components/Login')); 
+const Private = lazy(() => import('./Private/Private'));
 //esto de lazy es para que cargue el componente solo cuando se necesite, no al principio, recien cuando llame a Private lo va a cargar.
 
 function App() {  
 
   return (
     <div className="App" style={{ margin: "120px"}}>
-      {/* -------SE ME OCURRE HACER UN COMPONENTE SE TENGA EL SPINER PARA PONER ACA EN EL SUSPENSE Y EN DONDE CARGA */}
       <Suspense fallback={<Loading/>}> 
         <Provider store={store}>
           <BrowserRouter>
@@ -36,8 +34,10 @@ function App() {
                   <Route path='/' element={<Navigate to={PrivateRoutes.PRIVATE} />} /> 
 
                   <Route path={PublicRoutes.LOGIN} element={<Login/>} />
+                  {/* <Route path={PublicRoutes.CREATE_CLIENT} element={<useCreateClient />} /> */}
+                  {/* VER COMO SE COMPORTA EL DE ARRIBA CREATECLIENT creo que ya lo puedo sacas*/}
+
                   <Route path={PublicRoutes.VIEW_PRODUCT} element={<ProductosDetail />} />
-                  <Route path={PublicRoutes.CREATE_CLIENT} element={<CreateClient />} />
 
                   <Route element={<AuthGuard />}>
                     <Route path={`${PrivateRoutes.PRIVATE}/*`} element={<Private />} />
